fix: generate 010-wplogin.sh if it does not exist

+++ features/src/wordpress/setup-wordpress.sh
@@ -132,15 +132,17 @@ if ! wp core is-installed >/dev/null 2>&1; then
             --skip-themes
     fi
 
-    export WP_URL="${wp_url}"
-    # shellcheck disable=SC2016
-    envsubst '${WP_URL}' < /usr/share/wordpress/010-wplogin.tpl > "${HOME}/.local/share/vip-codespaces/login/010-wplogin.sh"
-
     wp user add-cap 1 view_query_monitor
 
     run-parts /var/lib/wordpress/postinstall.d
 else
     echo "WordPress already installed"
+fi
+
+if [ ! -f "${HOME}/.local/share/vip-codespaces/login/010-wplogin.sh" ]; then
+    export WP_URL="${wp_url}"
+    # shellcheck disable=SC2016
+    envsubst '${WP_URL}' < /usr/share/wordpress/010-wplogin.tpl > "${HOME}/.local/share/vip-codespaces/login/010-wplogin.sh"
 fi
 
 exit 0
